WIP: Move Draw to FloorRenderer
